Fix issue starting server after new merge
