+++ cv_generator/main/models/__init__.py
@@ -3,3 +3,4 @@ This module holds all models
 """
 from .skill import Skill
 from .working_experience import WorkingExperience
+from .personal_information import PersonalInformation
