+++ cv_generator/main/models/working_experience.py
@@ -15,18 +15,3 @@ class WorkingExperience(models.Model):
     
     def __str__(self):
         return self.company_name + ' | ' + self.title
-
-
-class PersonalInformation(models.Model):
-    name = models.CharField(max_length=25, blank=False)
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    email = models.EmailField(max_length=255, blank=False)
-    website_links = models.URLField(blank=True)
-    profile_photo = models.ImageField(upload_to="images/", blank=False)
-    phone_number = models.PositiveIntegerField(blank=False)
-    zip_code = models.PositiveIntegerField(blank=False)
-    street = models.CharField(max_length=50, blank=False)
-    city = models.CharField(max_length=50, blank=False)
-
-    def __str__(self):
-        return f'{self.name} | {self.email}'
