+++ cv_generator/main/views/__init__.py
@@ -1,0 +1,1 @@
+from .working_experience import WorkingExperienceView
